+++ starsky-node-client/thumbnail/IResults.ts
@@ -1,0 +1,6 @@
+
+export interface IResults {
+    searchQuery : string,
+    lastPageNumber: number,
+    fileHashList: Array<string>;
+}