+++ starsky/starskycore/Migrations/20190422190334_ImportIndexItemFilePath.cs
@@ -1,0 +1,22 @@
+﻿using Microsoft.EntityFrameworkCore.Migrations;
+
+namespace starskycore.Migrations
+{
+    public partial class ImportIndexItemFilePath : Migration
+    {
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.AddColumn<string>(
+                name: "FilePath",
+                table: "ImportIndex",
+                nullable: true);
+        }
+
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropColumn(
+                name: "FilePath",
+                table: "ImportIndex");
+        }
+    }
+}